#21: resolve import cycle
